Organize code a bit more
Just moving files around
